fix(zsh/core): Error if .zsh_history does not yet exist

+++ zsh/core.zsh
@@ -27,11 +27,13 @@ export __ZSH_HISTIGNORE="$__ZSH_HISTIGNORE|^$__ZSH_HISTIGNORE_EXACT$"
 export __ZSH_HISTIGNORE="$__ZSH_HISTIGNORE)"
 __ZSH_HISTIGNORE_RESULT="/tmp/.zsh_history_filtered_pid$$"
 # If a filter result is missing or old.
-if [ ! -f $__ZSH_HISTIGNORE_RESULT ] || [ test `find $__ZSH_HISTIGNORE_RESULT -mmin +1 > /dev/null 2>&1` ]; then
-  egrep -v "$__ZSH_HISTIGNORE" ~/.zsh_history > $__ZSH_HISTIGNORE_RESULT
-  echo "ZSH history cleanup: $(diff ~/.zsh_history $__ZSH_HISTIGNORE_RESULT | egrep "^<" | wc -l)"
-  cp $__ZSH_HISTIGNORE_RESULT ~/.zsh_history
-  rm -f $__ZSH_HISTIGNORE_RESULT
+if [ -f ~/.zsh_history ]; then
+  if [ ! -f $__ZSH_HISTIGNORE_RESULT ] || [ test `find $__ZSH_HISTIGNORE_RESULT -mmin +1 > /dev/null 2>&1` ]; then
+    egrep -v "$__ZSH_HISTIGNORE" ~/.zsh_history > $__ZSH_HISTIGNORE_RESULT
+    echo "ZSH history cleanup: $(diff ~/.zsh_history $__ZSH_HISTIGNORE_RESULT | egrep "^<" | wc -l)"
+    cp $__ZSH_HISTIGNORE_RESULT ~/.zsh_history
+    rm -f $__ZSH_HISTIGNORE_RESULT
+  fi
 fi
 
 export HISTFILE=~/.zsh_history
